move stripeElementsOptions to config
